+++ tests/test_admin.rs
@@ -2,8 +2,6 @@
 
 use backoff::{ExponentialBackoff, Operation};
 
-use futures::Future;
-
 use std::time::Duration;
 
 use rdkafka::admin::{
@@ -78,8 +76,8 @@ fn verify_delete(topic: &str) {
     .unwrap()
 }
 
-#[test]
-fn test_topics() {
+#[tokio::test]
+async fn test_topics() {
     let admin_client = create_admin_client();
     let opts = AdminOptions::new().operation_timeout(Some(Duration::from_secs(1)));
 
@@ -101,7 +99,7 @@ fn test_topics() {
 
         let res = admin_client
             .create_topics(&[topic1, topic2], &opts)
-            .wait()
+            .await
             .expect("topic creation failed");
         assert_eq!(res, &[Ok(name1.clone()), Ok(name2.clone())]);
 
@@ -124,7 +122,7 @@ fn test_topics() {
                 ],
                 &opts,
             )
-            .wait()
+            .await
             .expect("describe configs failed");
         let config1 = &res[0].as_ref().expect("describe configs failed on topic 1");
         let config2 = &res[1].as_ref().expect("describe configs failed on topic 2");
@@ -165,27 +163,28 @@ fn test_topics() {
         let partitions1 = NewPartitions::new(&name1, 5);
         let res = admin_client
             .create_partitions(&[partitions1], &opts)
-            .wait()
+            .await
             .expect("partition creation failed");
         assert_eq!(res, &[Ok(name1.clone())]);
 
-        let mut backoff = ExponentialBackoff::default();
-        backoff.max_elapsed_time = Some(Duration::from_secs(5));
-        (|| {
+        let mut tries = 0;
+        loop {
             let metadata = fetch_metadata(&name1);
             let topic = &metadata.topics()[0];
             let n = topic.partitions().len();
-            if n != 5 {
-                Err(format!("topic has {} partitions, but expected {}", n, 5))?;
+            if n == 5 {
+                break;
+            } else if tries >= 5 {
+                panic!("topic has {} partitions, but expected {}", n, 5);
+            } else {
+                tries += 1;
+                tokio::time::delay_for(Duration::from_secs(1)).await;
             }
-            Ok(())
-        })
-        .retry(&mut backoff)
-        .unwrap();
+        }
 
         let res = admin_client
             .delete_topics(&[&name1, &name2], &opts)
-            .wait()
+            .await
             .expect("topic deletion failed");
         assert_eq!(res, &[Ok(name1.clone()), Ok(name2.clone())]);
         verify_delete(&name1);
@@ -196,7 +195,7 @@ fn test_topics() {
     // creating topics.
     {
         let topic = NewTopic::new("ignored", 1, TopicReplication::Variable(&[&[0], &[0]]));
-        let res = admin_client.create_topics(&[topic], &opts).wait();
+        let res = admin_client.create_topics(&[topic], &opts).await;
         assert_eq!(
             Err(KafkaError::AdminOpCreation(
                 "replication configuration for topic 'ignored' assigns 2 partition(s), \
@@ -215,7 +214,7 @@ fn test_topics() {
 
         let res = admin_client
             .create_topics(vec![&topic], &opts)
-            .wait()
+            .await
             .expect("topic creation failed");
         assert_eq!(res, &[Ok(name.clone())]);
         let _ = fetch_metadata(&name);
@@ -223,7 +222,7 @@ fn test_topics() {
         // This partition specification is obviously garbage, and so trips
         // a client-side error.
         let partitions = NewPartitions::new(&name, 2).assign(&[&[0], &[0], &[0]]);
-        let res = admin_client.create_partitions(&[partitions], &opts).wait();
+        let res = admin_client.create_partitions(&[partitions], &opts).await;
         assert_eq!(
             res,
             Err(KafkaError::AdminOpCreation(format!(
@@ -237,7 +236,7 @@ fn test_topics() {
         let partitions = NewPartitions::new(&name, 2).assign(&[&[0], &[0]]);
         let res = admin_client
             .create_partitions(&[partitions], &opts)
-            .wait()
+            .await
             .expect("partition creation failed");
         assert_eq!(res, &[Err((name, RDKafkaError::InvalidReplicaAssignment))],);
     }
@@ -247,7 +246,7 @@ fn test_topics() {
         let name = rand_test_topic();
         let res = admin_client
             .delete_topics(&[&name], &opts)
-            .wait()
+            .await
             .expect("delete topics failed");
         assert_eq!(res, &[Err((name, RDKafkaError::UnknownTopicOrPartition))]);
     }
@@ -263,14 +262,14 @@ fn test_topics() {
 
         let res = admin_client
             .create_topics(vec![&topic1], &opts)
-            .wait()
+            .await
             .expect("topic creation failed");
         assert_eq!(res, &[Ok(name1.clone())]);
         let _ = fetch_metadata(&name1);
 
         let res = admin_client
             .create_topics(vec![&topic1, &topic2], &opts)
-            .wait()
+            .await
             .expect("topic creation failed");
         assert_eq!(
             res,
@@ -283,14 +282,14 @@ fn test_topics() {
 
         let res = admin_client
             .delete_topics(&[&name1], &opts)
-            .wait()
+            .await
             .expect("topic deletion failed");
         assert_eq!(res, &[Ok(name1.clone())]);
         verify_delete(&name1);
 
         let res = admin_client
             .delete_topics(&[&name2, &name1], &opts)
-            .wait()
+            .await
             .expect("topic deletion failed");
         assert_eq!(
             res,
@@ -302,15 +301,15 @@ fn test_topics() {
     }
 }
 
-#[test]
-fn test_configs() {
+#[tokio::test]
+async fn test_configs() {
     let admin_client = create_admin_client();
     let opts = AdminOptions::new();
     let broker = ResourceSpecifier::Broker(0);
 
     let res = admin_client
         .describe_configs(&[broker], &opts)
-        .wait()
+        .await
         .expect("describe configs failed");
     let config = &res[0].as_ref().expect("describe configs failed");
     let orig_val = config
@@ -323,16 +322,15 @@ fn test_configs() {
     let config = AlterConfig::new(broker).set("log.flush.interval.messages", "1234");
     let res = admin_client
         .alter_configs(&[config], &opts)
-        .wait()
+        .await
         .expect("alter configs failed");
     assert_eq!(res, &[Ok(OwnedResourceSpecifier::Broker(0))]);
 
-    let mut backoff = ExponentialBackoff::default();
-    backoff.max_elapsed_time = Some(Duration::from_secs(5));
-    (|| {
+    let mut tries = 0;
+    loop {
         let res = admin_client
             .describe_configs(&[broker], &opts)
-            .wait()
+            .await
             .expect("describe configs failed");
         let config = &res[0].as_ref().expect("describe configs failed");
         let entry = config.get("log.flush.interval.messages");
@@ -357,18 +355,20 @@ fn test_configs() {
                 is_sensitive: false,
             }
         };
-        if entry != Some(&expected_entry) {
-            Err(format!("{:?} != {:?}", entry, Some(&expected_entry)))?
+        if entry == Some(&expected_entry) {
+            break;
+        } else if tries >= 5 {
+            panic!("{:?} != {:?}", entry, Some(&expected_entry));
+        } else {
+            tries += 1;
+            tokio::time::delay_for(Duration::from_secs(1)).await;
         }
-        Ok(())
-    })
-    .retry(&mut backoff)
-    .unwrap();
+    }
 
     let config = AlterConfig::new(broker).set("log.flush.interval.ms", &orig_val);
     let res = admin_client
         .alter_configs(&[config], &opts)
-        .wait()
+        .await
         .expect("alter configs failed");
     assert_eq!(res, &[Ok(OwnedResourceSpecifier::Broker(0))]);
 }
@@ -376,8 +376,8 @@ fn test_configs() {
 // Tests whether each admin operation properly reports an error if the entire
 // request fails. The original implementations failed to check this, resulting
 // in confusing situations where a failed admin request would return Ok([]).
-#[test]
-fn test_event_errors() {
+#[tokio::test]
+async fn test_event_errors() {
     // Configure an admin client to target a Kafka server that doesn't exist,
     // then set an impossible timeout. This will ensure that every request fails
     // with an OperationTimedOut error, assuming, of course, that the request
@@ -388,31 +388,31 @@ fn test_event_errors() {
         .expect("admin client creation failed");
     let opts = AdminOptions::new().request_timeout(Some(Duration::from_nanos(1)));
 
-    let res = admin_client.create_topics(&[], &opts).wait();
+    let res = admin_client.create_topics(&[], &opts).await;
     assert_eq!(
         res,
         Err(KafkaError::AdminOp(RDKafkaError::OperationTimedOut))
     );
 
-    let res = admin_client.create_partitions(&[], &opts).wait();
+    let res = admin_client.create_partitions(&[], &opts).await;
     assert_eq!(
         res,
         Err(KafkaError::AdminOp(RDKafkaError::OperationTimedOut))
     );
 
-    let res = admin_client.delete_topics(&[], &opts).wait();
+    let res = admin_client.delete_topics(&[], &opts).await;
     assert_eq!(
         res,
         Err(KafkaError::AdminOp(RDKafkaError::OperationTimedOut))
     );
 
-    let res = admin_client.describe_configs(&[], &opts).wait();
+    let res = admin_client.describe_configs(&[], &opts).await;
     assert_eq!(
         res.err(),
         Some(KafkaError::AdminOp(RDKafkaError::OperationTimedOut))
     );
 
-    let res = admin_client.alter_configs(&[], &opts).wait();
+    let res = admin_client.alter_configs(&[], &opts).await;
     assert_eq!(
         res,
         Err(KafkaError::AdminOp(RDKafkaError::OperationTimedOut))